backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WhatsApp Integration
requests==2.31.0
httpx==0.25.2

# PDF Generation
reportlab==4.0.7
Pillow==10.1.0
arabic-reshaper==3.0.0
python-bidi==0.4.2

# Environment
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Date & Time
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
